+++ level1c4pps/tests/test_vgac2pps.py
@@ -80,7 +80,7 @@ class TestVgac2PPS(unittest.TestCase):
                        'add_offset': 273.15},
             'qual_flags':  {'dtype': 'int16', 'zlib': True,
                             'complevel': 4, '_FillValue': -32001.0},
-            'scanline_timestamps': {'dtype': 'float64',
+            'scanline_timestamps': {'dtype': 'int64',
                                     'zlib': True,
                                     'units': 'milliseconds since 1970-01-01',
                                     'complevel': 4,
